+++ src/lib/help.ts
@@ -1,9 +1,9 @@
-const pkg = { version: "0.0.11" };
+const pkg = { version: "0.0.12" };
 
 export function printHelp() {
   const messages = [
     "",
-    `Automatiqal CLI\x1b[31m v${pkg.version}\x1b[0m`,
+    `Automatiqal CLI\x1b[32;1m v${pkg.version}\x1b[0m`,
     "",
     "Usage: automatiqal [options]",
     "",
@@ -12,6 +12,7 @@ export function printHelp() {
     "--variables, -v     Location of the variable file (if needed)",
     "--json              Indicates that the run book file is in JSON format",
     "--output,    -o     Saves the result in the provided path",
+    "--connect,   -c     Test the connectivity. No tasks are ran",
     "--sample,    -s     Generate sample run book and variables files in the current folder",
     "--help,      -h     Shows this message",
     "",
@@ -21,12 +22,12 @@ export function printHelp() {
     "$ automatiqal --file ./deployment.json --json",
     "$ automatiqal --file ./deployment.yaml --output ./deployment_result.json",
     "",
-    "\x1b[33mIf you find Automatiqal CLI useful, please consider sponsoring the project:",
+    "\x1b[33;1mIf you find Automatiqal CLI useful, please consider sponsoring the project:",
     "https://github.com/informatiqal/automatiqal-cli\x1b[0m",
     "",
   ];
 
-  for (let message of messages) {
+  messages.forEach((message) => {
     console.log(message);
-  }
+  });
 }